Updated results of the Lift case study.
Changed some minor things in the other case studies.

Updated the naming of the labels of all the statechart transitions. Additionally for the discrete synthesis problems, the transitions are unique up to state of origin, target state and output, hence they are first identified and put in a set of transitions. Then after creating the transitions, the events or triggers associated with the transitions are associated with it in the statechart UML code.